compile on FreeBSD (Jun Kuriyama++)

+++ src/logic/manager/framework.cc
@@ -66,7 +66,7 @@ DISPATCH_CATCH(method, response)
 
 void framework::new_node(address addr, role_type id, shared_node n)
 {
-	LOG_WARN("new node ",id," ",addr);
+	LOG_WARN("new node ",(uint16_t)id," ",addr);
 	if(id == ROLE_MANAGER) {
 		if(addr != share->partner()) {
 			TLOGPACK("eP",3,
